% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BM010}
\alias{BM010}
\title{BM010: Ejendomspriser på boligmarkedet efter område, ejendomskategori og priser på realiserede handler}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100395 rows and 7 columns.}
\source{
\url{http://www.realkreditraadet.dk/Statistikker/Boligmarkedsstatistik/Data.aspx}
}
\usage{
BM010
}
\description{
Statistikken dækker fra K1 1992 til K1 2016.
}
\keyword{datasets}

