% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BM011}
\alias{BM011}
\title{BM011: Ejendomspriser på boligmarkedet efter postnr., ejendomskategori og priser på realiserede handler}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 516816 rows and 7 columns.}
\source{
\url{http://www.realkreditraadet.dk/Statistikker/Boligmarkedsstatistik/Data.aspx}
}
\usage{
BM011
}
\description{
Statistikken dækker fra K1 1992 til K1 2016.
}
\keyword{datasets}

