% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BM020}
\alias{BM020}
\title{BM020: Bevægelser på boligmarkedet efter område, ejendomskategori og bevægelsestype}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33810 rows and 7 columns.}
\source{
\url{http://www.realkreditraadet.dk/Statistikker/Boligmarkedsstatistik/Data.aspx}
}
\usage{
BM020
}
\description{
Statistikken dækker fra K1 2004 til K1 2016.
}
\keyword{datasets}

