% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BM030}
\alias{BM030}
\title{BM030: Salgstider på boligmarkedet efter område og ejendomskategori}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16906 rows and 6 columns.}
\source{
\url{http://www.realkreditraadet.dk/Statistikker/Boligmarkedsstatistik/Data.aspx}
}
\usage{
BM030
}
\description{
Statistikken dækker fra K1 2004 til K1 2016.
}
\keyword{datasets}

