% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BM031}
\alias{BM031}
\title{BM031: Salgstider på boligmarkedet efter postnr. og ejendomskategori}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 87024 rows and 6 columns.}
\source{
\url{http://www.realkreditraadet.dk/Statistikker/Boligmarkedsstatistik/Data.aspx}
}
\usage{
BM031
}
\description{
Statistikken dækker fra K1 2004 til K1 2016.
}
\keyword{datasets}

